fix incorrect opts in salt-ssh
When using salt-ssh with the local file client, the opts used by the
local client running salt-ssh get overwritten by the opts gathered from
the minion. This leads to, for example, the cachedir being created in
the wrong directory, if they are in different directories on either
machine.

This fixes that issue by detecting when the local file client is in use,
and if so, uses the master config instead of minion.